Address pr comment Clean and comment code
